Add attributes and restructuring
The attributes in fluent are a tool to group different messages together and are always bundled with a Term or a Message.

These can be used by adding at the end new Elements that start with a `.`.
```ftl
test = Something should
  be here and we are just testing.
  .test-attribute = Just add something to this by testing.
```

----

Restructuring of the way the different AST Elements are inherited from each other. Now all of the Messages inherit from the Attribute and the Term still inherits from the Message.