Refactor Code Strucutre
As sometimes things got a little bit messy the code is more structured into different classes, interfaces etc.

The biggest changes are:
StringSlice now has a new Util class that has all of the methods that don't really belong into StringSlice.

FluentPlaceable now does not have any subclasses anymore and is now an interface that many classes are inheriting.

FluentElement is now an interface instead of an abstract class.