Implement indentation mechanism
Fluent has a mechanism that removes all of the indentation in a message automatically so that the messages can look good and be multiline.

It is checking every line (except the first) and counting the whitespace so that it can get the smallest amount of whitespace.
This number of whitespace is then getting removed in the parsing step.