Purge StringSlice as attribute and minor cleanup
Removing all of the StringSlice references in the AST so that the garbage collector has the ability to remove this memory. It allows the library to use less memory in total.

The necessary references are now passed down as parameters in the methods.

----

Some minor cleanup in the way that there is no `valueOf` method anymore and with that FluentArgument is not generic anymore.

The FluentParser is now completely abstract so that it cannot be instantiated and the parse method now has the required parameter.