Fix: Newline bug between parameters and name
When a function or a term has some parameters they are handed over between (). But when these brackets are not on the same line these are not seen as their parameters and so it will be an error.
This has been fixed so that they will be seen as parameters of that function / term.