Add base pattern parsing system
The pattern parsing system is running through all of the available parsers and checks if they would work. If they would work they are checking if they would return a ParseError and if they don't a new element is added or the possible part is being skipped.

After that the TextElements are being added by always tracking the amount of characters that have been checked and if a newline happens or a parser returns a valid response it is getting cut out and added to the list.