Add Builder for FluentParserGroup
Add a builder and a new interface.
The builder can be accessed by using `ResourceParser.builder()`.

But for people that just want to use the library without modifying it, they should use the `ResourceParser.DEFAULT` attribute as this is static and does not need to be created each time it is being used.

Also add the base interface for all of the builders to ensure that all of the builders are similar.