Change result of infinite recursion
When infinite recursion was detected it previously has thrown a RuntimeException. This was better than just not handling infinite Recursion and creating a StackOverflow-Error.

Now it doesn't throw and Exception and it is just adding a `{???}` instead of the result of the term or message. This is more in line with the JavaScript implementation of Fluent.