Change `CannotPlaceable` error exception
Change the exception that is being thrown when the element itself is just being used as a normal placeable while it has the CannotPlaceable Interface implemented. This was previously a simple RuntimeException but this isn't really useful as nearly everything is a RuntimeException.
Now it is a FluentBuilderException to indicate that this is something that has been found out at the parsing stage.