Refactor: Change ImmutableArgumentList to EmptyArgumentList
Even though the implementing class for the empty argument list is an immutable argument list it does not completely reflect the usage for the actual argument list.

An immutable argument list would normally contain elements but those cannot be changed. This does not contain elements and these cannot be changed.