Move selection logic to `Selectable`s
The selection logic is now moved to the Selectables because this results in clearer code (before the method returned a function that checks every single element and now the elements are just checked by the Selectable).
It is also beneficial if this is being extended as now the developer has a lot more control over the logic behind the selection.